Migrate the remaining unit tests in UserRepositoryTest and delete the old file
